Enhanced chat functionality with recent message responses, updated RefreshToken model, improved JWT handling, and added robust exception handling. Updated gradle and security configurations for improved API and Swagger documentation support.

+++ src/main/kotlin/com/church/payload/refreshtoken/RefreshTokenResponse.kt
@@ -3,5 +3,5 @@ package com.church.payload.refreshtoken
 data class RefreshTokenResponse (
     val accessToken: String,
     val refreshToken: String,
-    val expiresIn: Int? = null,
+    val expiresIn: Long? = null,
 )
